Fixes #73: replace call to get_role_users with sql.
The call to get_role_users was triggering a call to 'debugging'
in get_records_sql because the call could return more than one
instance of the same user (u.id) if that user had >1 role assignments
in a course.
The debugging message generated is:
"Did you remember to make the first column something unique in your
call to get_records? Duplicate value '2736' found in column 'id'. "